# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Station Satcom Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of admin users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminUserDetail" }
    post:
      operationId: createAdminUser
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAdminUserBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserDetail" }
  /admin/users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    patch:
      operationId: updateAdminUser
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAdminUserBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserDetail" }
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
  /admin/users/{id}/reset-password:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    post:
      operationId: resetAdminUserPassword
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResetPasswordBody" }
      responses:
        "200":
          description: Reset
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    MessageResponse:
      type: object
      properties:
        message: { type: string }
      required: [message]
    LoginBody:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string }
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/AdminUser"
      required: [token, user]
    AdminUser:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [owner, admin, viewer] }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, name, email, role, createdAt]
    AdminUserDetail:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [owner, admin, viewer] }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        lockedUntil: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, email, role, createdAt, updatedAt]
    CreateAdminUserBody:
      type: object
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        role: { type: string, enum: [owner, admin, viewer] }
      required: [name, email, password]
    UpdateAdminUserBody:
      type: object
      properties:
        name: { type: string }
        role: { type: string, enum: [owner, admin, viewer] }
        unlock: { type: boolean }
    ResetPasswordBody:
      type: object
      properties:
        newPassword: { type: string }
      required: [newPassword]
...
    ChangePasswordBody:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
      required: [currentPassword, newPassword]
...